finish v1

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -109,8 +109,11 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
-        $category = Category::find($id);
-        $category->delete();
-        return redirect(route('category.index'))->with(['success' => 'Kategori Dihapus!']);
+        $category = Category::withCount(['product'])->find($id);
+        if ($category->product_count == null) {
+            $category->delete();
+            return redirect()->route('category.index')->with(['success' => 'Kategori Dihapus!']);
+        }
+        return redirect()->route('category.index')->with(['warning' => 'Kategori Ini berelasi dengan data produk !']);
     }
 }
